requirements.txt
pandas
numpy
matplotlib
plotly
requests
python-dateutil
pytz
skyfield
jplephem
scikit-learn
